backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
numpy
tensorflow-macos
tensorflow-metal
opencv-python


